Move functions in chianti_kev_lines into more sensible modules.
