+++ sunxspex/utils/__init__.py
@@ -2,3 +2,5 @@
 
 # This sub-module is destined for common non-package specific utility
 # functions.
+
+from . utils import *
